+++ lib/smartgen/engines.rb
@@ -1,3 +1,4 @@
 require File.expand_path('engines/base', File.dirname(__FILE__))
 require File.expand_path('engines/textile', File.dirname(__FILE__))
 require File.expand_path('engines/markdown', File.dirname(__FILE__))
+require File.expand_path('engines/erb', File.dirname(__FILE__))