+++ lib/smartgen/engines/markdown.rb
@@ -5,7 +5,7 @@ module Smartgen
     # Processes markdown files, supporting both '.md' and '.markdown' extensions.
     class Markdown < Base
       protected
-        def parse(body)
+        def parse(body, metadata)
           BlueCloth.new(body).to_html
         end